Move whisper pipeline bindings to a separate file
